Crate eq_primitives

Equilibrium Primitives Pallet

Equilibrium's Primitives Pallet is a module containing basic data types and interfaces for Equilibrium pallets

Re-exports

pub use crate::signed_balance::SignedBalance;

Modules

currency
Module to set list of available currencies and implement functions for currency operations

signed_balance
Implementation of stored positive and negative balances

Structs

TotalAggregates
Aggregated USD values of positive (collateral) and negative (debt) balances for a group of users

Enums

InterestRateError
Errors during fee calculation

TransferReason
Describe the reason for the transfer.

UserGroup
User groups used in Equilibrium substrate

Traits

AccountGetter
Supports getting account id

Aggregates
When implemented, pallet can work with TotalAggregates

FeeManager
Generic interface for calculating and charging fees. Used in Treasury and Bailsman pallets